Se ha implementado el uso de mensajes ante cualquier interaccion del usuario con el sistema

+++ frontend/src/app/components/equipo/equipo.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { SEquipoService } from 'src/app/services/s-equipo.service';
 import { EquipoInterface, PaisInterface } from 'src/app/models/equipo-interface';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-equipo',
@@ -28,42 +29,78 @@ export class EquipoComponent implements OnInit {
   guardarEquipo(){
     this.fecha_fundacion = this.fecha_fundacion.split("-").reverse().join("/");
     this.sequipoervice.InsertEquipos(this.nombre, this.fecha_fundacion, this.link_fotografia, this.id_pais)
-    .subscribe(
-      (response) => {
+    .subscribe((res:any) => {
+      if (res['response']) {
+        Swal.fire({
+          icon: 'success',
+          title: 'Exito',
+          text: res['msg']
+        });
         this.cargarTabla();
         this.limpiarDatos();
-      },
-      (error) => {
-        console.log("El equipo ya existe en la base de datos");
+      }else{
+        Swal.fire({
+          icon: 'error',
+          title: 'Oops...',
+          text: res['msg']
+        });
       }
-    );
+    });
   }
 
   eliminarEquipo(id_equipo:number){
-    var hola = this.sequipoervice.DeleteEquipo(id_equipo)
-    .subscribe(
-      (response) => {
-        this.cargarTabla();
-        this.limpiarDatos();
-      },
-      (error) => {
-        console.log("El dato que desea eliminar esta siendo utilizado");
+    Swal.fire({
+      title: '¿Esta seguro?',
+      text: "No podras revertir esto",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Eliminar!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.sequipoervice.DeleteEquipo(id_equipo)
+        .subscribe((res:any) => {
+          if (res['response']) {
+            Swal.fire({
+              icon: 'success',
+              title: 'Exito',
+              text: res['msg']
+            });
+            this.cargarTabla();
+            this.limpiarDatos();
+          }else{
+            Swal.fire({
+              icon: 'error',
+              title: 'Oops...',
+              text: res['msg']
+            });
+          }
+        });  
       }
-    );    
+    });      
   }
 
   actualizarEquipo(){
     this.fecha_fundacion = this.fecha_fundacion.split("-").reverse().join("/");
     this.sequipoervice.UpdateEquipos(this.id_equipo, this.nombre, this.fecha_fundacion, this.link_fotografia, this.id_pais)
-    .subscribe(
-      (response) => {
+    .subscribe((res:any) => {
+      if (res['response']) {
+        Swal.fire({
+          icon: 'success',
+          title: 'Exito',
+          text: res['msg']
+        });
         this.cargarTabla();
         this.limpiarDatos();
-      },
-      (error) => {
-        console.log("Los cambios que quiere realizar ya existen en el sistema");
+      }else{
+        Swal.fire({
+          icon: 'error',
+          title: 'Oops...',
+          text: res['msg']
+        });
       }
-    );
+    });
   }
 
   setearInterfaz(id_equipo:number, nombre:string, fecha_fundacion:string, link_fotografia:string, id_pais:number){
